+++ src/unitorch/models/peft/diffusers/modeling_stable_flux.py
@@ -104,16 +104,18 @@ class GenericStableFluxLoraModel(GenericPeftModel, QuantizationMixin):
             config_dict.update({"in_channels": in_channels})
         if out_channels is not None:
             config_dict.update({"out_channels": out_channels})
-        self.transformer = FluxTransformer2DModel.from_config(config_dict)
+        self.transformer = FluxTransformer2DModel.from_config(config_dict).to(
+            torch.bfloat16
+        )
 
         text_config = CLIPTextConfig.from_json_file(text_config_path)
-        self.text = CLIPTextModel(text_config)
+        self.text = CLIPTextModel(text_config).to(torch.bfloat16)
 
         text_config2 = CLIPTextConfig.from_json_file(text2_config_path)
-        self.text2 = T5EncoderModel(text_config2)
+        self.text2 = T5EncoderModel(text_config2).to(torch.bfloat16)
 
         vae_config_dict = json.load(open(vae_config_path))
-        self.vae = AutoencoderKL.from_config(vae_config_dict)
+        self.vae = AutoencoderKL.from_config(vae_config_dict).to(torch.bfloat16)
 
         if isinstance(controlnet_configs_path, list):
             if len(controlnet_configs_path) == 0:
@@ -126,7 +128,9 @@ class GenericStableFluxLoraModel(GenericPeftModel, QuantizationMixin):
             for controlnet_config_path in controlnet_configs_path:
                 controlnet_config_dict = json.load(open(controlnet_config_path))
                 controlnets.append(
-                    FluxControlNetModel.from_config(controlnet_config_dict)
+                    FluxControlNetModel.from_config(controlnet_config_dict).to(
+                        torch.bfloat16
+                    )
                 )
             self.num_controlnets = len(controlnets)
             self.controlnet = FluxMultiControlNetModel(
@@ -134,7 +138,9 @@ class GenericStableFluxLoraModel(GenericPeftModel, QuantizationMixin):
             )
         elif isinstance(controlnet_configs_path, str):
             controlnet_config_dict = json.load(open(controlnet_configs_path))
-            self.controlnet = FluxControlNetModel.from_config(controlnet_config_dict)
+            self.controlnet = FluxControlNetModel.from_config(
+                controlnet_config_dict
+            ).to(torch.bfloat16)
             self.num_controlnets = 1
         else:
             self.controlnet = None
